flask-backend/requirement.txt
pandas 
scikit-learn==1.1.1 
mysql-connector-python
flask_mysqldb